pyproject.toml
[project]
name = "photo-manager"
version = "0.1.0"
description = "Local-only, folder-based photo manager with face recognition"
authors = [{ name = "PhotoApp Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyside6>=6.6",
  "pillow>=10.0",
  "insightface>=0.7.3",
  "numpy>=1.24",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "mypy>=1.8",
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ["src/tests"]
